finish member and project testing
